WIP 8/12

+++ addons/sourcemod/scripting/zombie_escape/filenetwork.sp
@@ -17,47 +17,100 @@
 #pragma semicolon 1
 #pragma newdecls required
 
+#tryinclude <filenetwork>
+
+#define FILENET_LIBRARY	"filenetwork"
+
+#if defined _filenetwork_included
+static bool Loaded;
+
 static bool StartedQueue[MAXTF2PLAYERS];
 static bool Downloading[MAXTF2PLAYERS];
 
 static ArrayList SoundList;
+#endif
 
-void FileNetwork_PluginStart()
+void FileNet_PluginStatus()
 {
-	SoundList = new ArrayList(ByteCountToCells(PLATFORM_MAX_PATH));
+	#if defined _filenetwork_included
+	PrintToServer("File-Network: %s", Loaded ? "Running" : "Library not running");
+	#else
+	PrintToServer("File-Network: Compiled without include \"filenetwork\"");
+	#endif
 }
 
-void FileNetwork_MapEnd()
+void FileNet_PluginStart()
 {
-	for(int i; i < sizeof(SoundLevel); i++)
-	{
-		SoundLevel[i] = 0;
-	}
+	#if defined _filenetwork_included
+	SoundList = new ArrayList(ByteCountToCells(PLATFORM_MAX_PATH));
+	#endif
+}
 
+stock void FileNet_LibraryAdded(const char[] name)
+{
+	#if defined _filenetwork_included
+	if(!Loaded && StrEqual(name, FILENET_LIBRARY))
+	{
+		Loaded = true;
+
+		for(int client = 1; client <= MaxClients; client++)
+		{
+			if(StartedQueue[client] && !Downloading[client])
+				SendNextFile(client);
+		}
+	}
+	#endif
+}
+
+stock void FileNet_LibraryRemoved(const char[] name)
+{
+	#if defined _filenetwork_included
+	if(Loaded && StrEqual(name, FILENET_LIBRARY))
+		Loaded = false;
+	#endif
+}
+
+void FileNet_MapEnd()
+{
+	#if defined _filenetwork_included
 	delete SoundList;
 
-	FileNetwork_PluginStart();
+	FileNet_PluginStart();
+	#endif
 }
 
-void FileNetwork_ClientPutInServer(int client)
+void FileNet_ClientPutInServer(int client)
 {
-	FileNetwork_ClientDisconnect(client);
+	#if defined _filenetwork_included
+	FileNet_ClientDisconnect(client);
 	SendNextFile(client);
+	#else
+	QueryClientConVar(client, "cl_allowdownload", FileNet_QueryAllowDownload);
+	#endif
 }
 
-void FileNetwork_ClientDisconnect(int client)
+stock void FileNet_ClientDisconnect(int client)
 {
+	#if defined _filenetwork_included
 	StartedQueue[client] = false;
 	Downloading[client] = false;
-	Client(client).SoundLevel = 0;
+	#endif
 }
 
-int PrecacheSoundCustom(const char[] sound)
+int FileNet_PrecacheSound(const char[] sound)
 {
+	#if defined _filenetwork_included
 	PrecacheSound(sound);
-	return AddSoundFile(buffer);
+	return AddSoundFile(sound);
+	#else
+	char download[PLATFORM_MAX_PATH];
+	FormatEx(download, sizeof(download), "sound/%s", sound[sound[0] == '#' ? 1 : 0]);
+	AddFileToDownloadsTable(download);
+	return 0;
+	#endif
 }
 
+#if defined _filenetwork_included
 static int AddSoundFile(const char[] sound)
 {
 	int index = SoundList.FindString(sound);
@@ -88,45 +141,43 @@ static void SendNextFile(int client)
 
 	StartedQueue[client] = true;
 	
-	static char download[PLATFORM_MAX_PATH];
-	DataPack pack;
-
-	if(Client(client).SoundLevel < SoundList.Length)
+	if(Loaded)
 	{
-		SoundList.GetString(Client(client).SoundLevel, download, sizeof(download));
-		Format(download, sizeof(download), "sound/%s", download[download[0] == '#' ? 1 : 0]);
-		
-		pack = new DataPack();
-	}
+		static char download[PLATFORM_MAX_PATH];
+		DataPack pack;
 
-	if(pack)
-	{
-		Downloading[client] = true;
-
-		pack.WriteString(download);
-		
-		static char filecheck[PLATFORM_MAX_PATH];
-		FormatFileCheck(download, client, filecheck, sizeof(filecheck));
-#if !defined UseDownloadTable
-		FileNet_RequestFile(client, filecheck, FileNetwork_RequestResults, pack);
-#endif
-		if(!DeleteFile(filecheck, true))	// There has been some cases where we still have a file (Eg. plugin unload)
+		if(Client(client).SoundLevel < SoundList.Length)
 		{
-			Format(filecheck, sizeof(filecheck), "download/%s", filecheck);
-			DeleteFile(filecheck);
+			SoundList.GetString(Client(client).SoundLevel, download, sizeof(download));
+			Format(download, sizeof(download), "sound/%s", download[download[0] == '#' ? 1 : 0]);
+			
+			pack = new DataPack();
 		}
-	}
-	else
-	{
-		Downloading[client] = false;
 
-		PrintToConsole(client, "---");
-		PrintToConsole(client, "[ZR/RPG] Finished Downloading/Verifying Files! You will hear and see everything as intended now.");
-		PrintToConsole(client, "---");
+		if(pack)
+		{
+			Downloading[client] = true;
+
+			pack.WriteString(download);
+			
+			static char filecheck[PLATFORM_MAX_PATH];
+			FormatFileCheck(download, client, filecheck, sizeof(filecheck));
+			FileNet_RequestFile(client, filecheck, FileNet_RequestResults, pack);
+
+			if(!DeleteFile(filecheck, true))	// There has been some cases where we still have a file (Eg. plugin unload)
+			{
+				Format(filecheck, sizeof(filecheck), "download/%s", filecheck);
+				DeleteFile(filecheck);
+			}
+		}
+		else
+		{
+			Downloading[client] = false;
+		}
 	}
 }
 
-public void FileNetwork_RequestResults(int client, const char[] file, int id, bool success, DataPack pack)
+public void FileNet_RequestResults(int client, const char[] file, int id, bool success, DataPack pack)
 {
 	// If not found, send the actual file
 
@@ -155,8 +206,8 @@ public void FileNetwork_RequestResults(int client, const char[] file, int id, bo
 		else
 		{
 			// So the client doesn't freak out about existing CreateFragmentsFromFile spam
-			PrintToConsole(client, "[ZR/RPG/ZE] Downloading '%s'", download);
-			if(FileNet_SendFile(client, download, FileNetwork_SendResults, pack))
+			PrintToConsole(client, "[ZE] Downloading '%s'", download);
+			if(FileNet_SendFile(client, download, FileNet_SendResults, pack))
 				return;
 			
 			LogError("Failed to queue file \"%s\" to client", download);
@@ -166,7 +217,7 @@ public void FileNetwork_RequestResults(int client, const char[] file, int id, bo
 	delete pack;
 }
 
-public void FileNetwork_SendResults(int client, const char[] file, bool success, DataPack pack)
+public void FileNet_SendResults(int client, const char[] file, bool success, DataPack pack)
 {
 	// When done, send a dummy file and the next file in queue
 	
@@ -181,7 +232,7 @@ public void FileNetwork_SendResults(int client, const char[] file, bool success,
 			filec.WriteLine("Used for file checks for ZR/RPG/ZE");
 			filec.Close();
 
-			if(!FileNet_SendFile(client, filecheck, FileNetwork_SendFileCheck))
+			if(!FileNet_SendFile(client, filecheck, FileNet_SendFileCheck))
 			{
 				LogError("Failed to queue file \"%s\" to client", filecheck);
 				if(!DeleteFile(filecheck))
@@ -200,7 +251,7 @@ public void FileNetwork_SendResults(int client, const char[] file, bool success,
 	delete pack;
 }
 
-public void FileNetwork_SendFileCheck(int client, const char[] file, bool success)
+public void FileNet_SendFileCheck(int client, const char[] file, bool success)
 {
 	// Delete the dummy file left over
 
@@ -210,3 +261,30 @@ public void FileNetwork_SendFileCheck(int client, const char[] file, bool succes
 	if(!DeleteFile(file))
 		LogError("Failed to delete file \"%s\"", file);
 }
+
+#else
+
+public void FileNet_QueryAllowDownload(QueryCookie cookie, int client, ConVarQueryResult result, const char[] cvarName, const char[] cvarValue, any value)
+{
+	if(IsClientInGame(client) && result == ConVarQuery_Okay && !StringToInt(cvarValue))
+	{
+		QueryClientConVar(client, "sv_allowupload", FileNet_QueryAllowUpload);
+	}
+}
+
+public void FileNet_QueryAllowUpload(QueryCookie cookie, int client, ConVarQueryResult result, const char[] cvarName, const char[] cvarValue, any value)
+{
+	if(IsClientInGame(client) && result == ConVarQuery_Okay && !StringToInt(cvarValue))
+	{
+		QueryClientConVar(client, "cl_downloadfilter", FileNet_QueryFilter);
+	}
+}
+
+public void FileNet_QueryFilter(QueryCookie cookie, int client, ConVarQueryResult result, const char[] cvarName, const char[] cvarValue, any value)
+{
+	if(IsClientInGame(client) && result == ConVarQuery_Okay && StrContains("all", cvarValue) != -1)
+	{
+		Client(client).SoundLevel = 1;
+	}
+}
+#endif
